% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{addgrid}
\alias{addgrid}
\title{addgrid adds a grid to a canvas within a diagram}
\usage{
addgrid(left = 0, right = 100, bottom = 0, top = 100, step = 20)
}
\arguments{
\item{left}{defines lefthand edge of rectangle default = 0}

\item{right}{defines right-hand edge of rect default = 100}

\item{bottom}{defines bottom edge of rectangle angle default = 0}

\item{top}{defines top edge of rectangle default = 100}

\item{step}{how often to draw a grid line, default = 20}
}
\value{
nothing but it does add a grid to a canvas
}
\description{
addgrid generates a grid to a canvas to act as a guide for
    when adding rectangles, etc.
}
\examples{
\dontrun{
 plotprep(width=8,height=5)
 center <- makecanvas(xstart=0,xfinish=100,ystart=0,yfinish=100,addbox=TRUE)
 addgrid()  # remove when ready to make final plot
}
}
