% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{anglearrow}
\alias{anglearrow}
\title{anglearrow draws an angled arrow either left-right or right-left}
\usage{
anglearrow(
  boxa,
  boxb,
  location = "leftright",
  len = 0.1,
  ang = 30,
  code = 2,
  col = "black",
  lwd = 3
)
}
\arguments{
\item{boxa}{the first box, output from textbox}

\item{boxb}{the second box, output from textbox}

\item{location}{default = 'leftright', meaning left-above to right-below, 
'rightleft' implies from right-above to left-below. Currently these two
options are the only available.}

\item{len}{length of the arrow head, default = 0.1}

\item{ang}{angle of the arrow head, default = 30 degrees}

\item{code}{type of arrow head, default = 2, = arrow pointing to second box}

\item{col}{the colour of the lines definning the box, default = black}

\item{lwd}{the linewidth of hte arrow default = 3}
}
\value{
nothing but it does draw an angled arrow
}
\description{
anglearrow is used when drawing flowcharts and similar where
    instead of a straight arrow from top to bottom or left to right, one 
    textbox is to the left and below another or to the right and above the 
    other. This means there is a need to draw a line followed by an arrow.
}
\examples{
plotprep(width=8,height=7)
canvas <- makecanvas()
box2 <- textbox(x1=6,x2=53,y1=75,y2=65,col="black",fill=NA,lwd=1,
                txt=c("Includes an If Statement"),cex=1.5) 
box3 <- textbox(x1=65,x2=90,y1=62,y2=57,col="black",fill=NA,lwd=1,
                txt=c("An Alternative"),cex=1.1)
box4 <- textbox(x1=6,x2=53,y1=55,y2=46,col="black",fill=NA,lwd=1,
                txt=c("Hello World","A second line",
                      "A final third line"),cex=1.1)
arrows(x0=box2[[2]][1],y0=min(box2[[1]][,2]),x1=box4[[2]][1],
       y1=max(box4[[1]][,2]),length=0.1,angle=30,code=2,col="black",lwd=3)
anglearrow(box2,box3,location="leftright")
anglearrow(box3,box4,location="rightleft")
}
