% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{drawarrow}
\alias{drawarrow}
\title{drawarrow draws an angled arrow either left-right or right-left}
\usage{
drawarrow(
  box1,
  box2,
  location = "down",
  len = 0.1,
  ang = 30,
  code = 2,
  col = "black",
  lwd = 3
)
}
\arguments{
\item{box1}{the first box, output from textbox, from which an arrow starts}

\item{box2}{the second box, output from textbox, to which an arrow arrives}

\item{location}{default = 'down', meaning draw an arrow starting from the
bottom of box1 down to the top of box2. Viable options include: 'up', 
'down','left','right','upright','upleft','downright', and 'downleft'}

\item{len}{length of the arrow head, default = 0.1}

\item{ang}{angle of the arrow head, default = 30 degrees}

\item{code}{type of arrow head, default = 2, = arrow pointing to second box}

\item{col}{the colour of the lines defining the box, default = black}

\item{lwd}{the line width of the arrow default = 3}
}
\value{
nothing but it does draw an angled arrow
}
\description{
drawarrow is used when drawing flowcharts and similar. The idea
    is to describe how two boxes, box1 and box2, are joined, where box1 is 
    the first input and box2 the second. So if the third input, 'location', 
    is 'up' this would imply drawing an arrow from box1 up to box2, if 
    location is 'down' then an arrow from box1 down to box2. Similarly with
    'left' and 'right'. We sometimes need bent arrows to join 2 boxes that 
    are on a diagonal to each other. So, 'upright' would imply an arrow that 
    first rises up from box1 and then turns right to join box2, whichis up 
    and to the right of box1. Similarly, 'upleft' rises from box1 and turns 
    left to join box2. The possibilities include: 'up', 'down', 'left',
    'right', 'upright', 'upleft', 'downright', and 'downleft'. See the 
    example code.
}
\examples{
plotprep(width=8,height=7)
canvas <- makecanvas()
boxa <- textbox(x1=6,x2=53,y1=75,y2=65,col="black",fill=NA,lwd=1,
                txt=c("a: Includes an If Statement"),cex=1.5)
boxb <- textbox(x1=65,x2=90,y1=62,y2=57,col="black",fill=NA,lwd=1,
                txt=c("b: An Alternative"),cex=1.1)
boxc <- textbox(x1=6,x2=53,y1=55,y2=46,col="black",fill=NA,lwd=1,
                txt=c("c: Hello World","A second line",
                      "A final third line"),cex=1.1)
drawarrow(box1=boxa,box2=boxc,location="down")
drawarrow(box1=boxa,box2=boxb,location="rightdown")
drawarrow(box1=boxb,box2=boxc,location="downleft")
}
\seealso{
{
   \link{makecanvas}, \link{textbox}, \link{flowcharttemplate}
}
}
