% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{flowcharttemplate}
\alias{flowcharttemplate}
\title{flowcharttemplate write a flowchart template to the console}
\usage{
flowcharttemplate()
}
\value{
R code to the console using hplot functions to draw a flowchart
}
\description{
flowcharttemplate write a flowchart template to the console
    which uses a set of hplot functions to create a canvas and text boxes
    with arrows between them. The idea is to provide a working example that
    can be modified while illustrating the use of the hplot functions.
}
\examples{
flowcharttemplate()
}
