% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{makecanvas}
\alias{makecanvas}
\title{makecanvas sets up a plotting area ready for the flowchart}
\usage{
makecanvas(start = 0, finish = 100, addgrid = FALSE, addbox = FALSE)
}
\arguments{
\item{start}{x- and y-origin value defaults = 0}

\item{finish}{maximum of x- and y-axes defaults = 100}

\item{addgrid}{should a guiding grid and labeled axes be drawn to aid the
placing of graphical objects on the canvas. Once completed remove the 
grid and re-plot. default = FALSE}

\item{addbox}{should a box be drawn around the canvas, default = FALSE. This
may be desirable after drafting a figure using addgrid.}
}
\value{
plots an empty graph ready for polygons and text, returns a 
    list defining the objects on the canvas
}
\description{
makecanvas sets up a plotting areas ready for a flowchart
    made up of shapes, circles, polygons, rectangles, text, and arrows.
    A plotprep width of 8 and height of 10 will fit an A4 page well.
}
\examples{
  canvas=makecanvas(start=0,finish=100)
  polygon(makevx(2,27),makevy(90,6),col=0,lwd=1,border=1)
}
\seealso{
{
   \link{drawarrow}, \link{textbox}, \link{flowcharttemplate},
   \link{plotprep}
}
}
