% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{makeoblong}
\alias{makeoblong}
\title{makeoblong draws a rectangle once a canvas is available}
\usage{
makeoblong(x1, x2, y1, y2, col = "black", lwd = 1, ...)
}
\arguments{
\item{x1}{defines lefthand edge of rectangle}

\item{x2}{defines right-hand edge or rectangle}

\item{y1}{defines bottom edge of rectangle}

\item{y2}{defines top edge of rectangle}

\item{col}{colour of line. default="black"}

\item{lwd}{the width of the line, default=1}

\item{...}{to allow for other graphics commands such as lty=2, etc}
}
\value{
an invisible vector denoting the center (x,y) of the rectangle
}
\description{
makeoblong draws a rectangle after makecanvas has been called
}
\examples{
\dontrun{
   makecanvas(xstart=0,xfinish=100,ystart=0,yfinish=100)
   makeoblong(x1=20,x2=80,y1=25,y2=75,col="red",lwd=2)
}
}
