% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{makerect}
\alias{makerect}
\title{makerect draws a rectangle once a plot is available}
\usage{
makerect(left, right, bottom, top, linecol = "grey", lwd = 1, console = FALSE)
}
\arguments{
\item{left}{defines lefthand edge of rectangle}

\item{right}{defines right-hand edge of rect}

\item{bottom}{defines bottom edge of rectangle angle}

\item{top}{defines top edge of rectangle}

\item{linecol}{colour of line. default="grey"}

\item{lwd}{the width of the line, default=1}

\item{console}{should the centerx and centery bt sent ot console, 
default=FALSE}
}
\value{
invisibly a vector denoting the center (x,y) of the rectangle
}
\description{
makerect draws a rectangle after canvas has been called
}
\examples{
\dontrun{
   canvas(ystart=50,yfinish=93.5)
   makerect(left=2,xinc=27,top=90,yinc=6)
}
}
