% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{makevx}
\alias{makevx}
\title{makevx make an x values vector for plotting oblongs}
\usage{
makevx(x1, x2)
}
\arguments{
\item{x1}{x-value for the left-hand edge of a rectangle}

\item{x2}{x value for the the right-hand edge}
}
\value{
a vector of x-values
}
\description{
makevx takes the left x value of a rectangle and the
    right x value and defines a vector describing the four
    vertices of the rectangle bottomleft, bottomright, topright,
    topleftleft, bottomleft. when matched with makevy generates the
    descriptor for a complete rectangle or oblong.
}
\examples{
\dontrun{
 plot(0:100,seq(0,100),type="n",xaxt="n",yaxt="n",
 xlab="",ylab="", bty="n")
 lines(makevx(20,80),makevy(20,80),col=1,lwd=2)
}
}
