% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{makevy}
\alias{makevy}
\title{makevy make a y values vector}
\usage{
makevy(y1, y2)
}
\arguments{
\item{y1}{y-value for the top edge of a rectangle}

\item{y2}{the y-increment subtracted from init to define the lower edge}
}
\value{
a vector of y-values
}
\description{
makevy takes the bottom y value of a rectangle and the
    top y value and defines a vector describing the four
    vertices of the rectangle bottomleft, bottomright, topright, topleft,
    bottomleft. When matched with makevx generates the
    descriptor for a complete rectangle.
}
\examples{
\dontrun{
 canvas(ystart=50,yfinish=93.5)
 polygon(makevx(2,27),makevy(90,6),col=0,lwd=1,border=1)
}
}
