% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{panel.hist}
\alias{panel.hist}
\title{panel.hist a function to modify the pairs function}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{a data.frame or matrix of variables that will be plotted
against each other in a pairs plot}

\item{...}{the other parameters for the function}
}
\value{
nothing but it does add histograms to the diagonals
}
\description{
panel.hist is a function lifted directly from the help
    for the 'pairs' function. It enables histograms of each variable
    to be included in the diagonal plots inside the pairs plot. It is
    only included here as a convenience.
}
\examples{
print("wait on data sets")
# syntax
# pairs(x,...,diag.panel=panel.hist)
}
