% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{plotnull}
\alias{plotnull}
\title{plotnull generates an empty plot when one is needed}
\usage{
plotnull(msg = "", cex = 1, x = 10, y = 45)
}
\arguments{
\item{msg}{a message to be printed in the middle of the empty plot.}

\item{cex}{what font size to use in any message, default = 1.0}

\item{x}{the x location (0 - 100) of the msg default = 10}

\item{y}{the y location (0 - 100) of left side of message, default = 45}
}
\value{
nothing but it does generate a plot
}
\description{
plotnull, there are often circumstances, for example, when
    plotting up results from each year and each SAU, where there will be
    combinations of year and SAU that have no data, but to avoid a problem
    with the plotting it is necessary to generate an empty plot. By using
    x=0:100 and y= 0:100 this is equivalent to a canvas of 0:100,0:100
}
\examples{
plotnull("An empty plot",cex=1.5)
}
