% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{pltregcol}
\alias{pltregcol}
\title{pltregcol colours the plot region of a base R plot using panel.first}
\usage{
pltregcol(col = "grey95", gridcol = "darkgrey")
}
\arguments{
\item{col}{what grey colour to use to infill, default = 'grey95'. Can also
just enter a number 15-99, which will generate different shades of grey
in case grey95 is too light or too dark}

\item{gridcol}{what colour to use for the grid added to the plot, default=
'darkgrey'}
}
\value{
nothing but it does infill the plot region of a plot
}
\description{
pltregcol is used to fill in the plot area with a neutral grey 
    colour to clarify the use of some of the fainter colours, especially
    when using faint transparent colours. The solution came from a
    stackoverflow discussion. It is used as the argument for the panel.first
    function.
}
\examples{
plot(x=2,y=2,panel.first=pltregcol())
}
