% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{putcircle}
\alias{putcircle}
\title{putcircle draws a circle with a given origin and radius}
\usage{
putcircle(
  origx = 50,
  origy = 50,
  radius = 10,
  col = 1,
  lwd = 1,
  fill = NA,
  ...
)
}
\arguments{
\item{origx}{the final x origin}

\item{origy}{the final y origin}

\item{radius}{the radius of the circle}

\item{col}{the col of the circle}

\item{lwd}{the line width of the circle}

\item{fill}{should the circle be filled with colour? default=NA = empty}

\item{...}{to allow for other graphics commands such as lty=2, etc}
}
\value{
the matrix of x and y values invisibly
}
\description{
putcircle provides the means of drawing a circle of a given
    radius and origin within a diagram ready for the addition of text.
}
\examples{
  makecanvas()
  putcircle(origx=35,origy=70,radius=30,lwd=2,col=1)
  putcircle(origx=65,origy=60,radius=30,lwd=2,col=2)
  putcircle(origx=45,origy=40,radius=30,lwd=2,col=4)
}
