% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{putoblong}
\alias{putoblong}
\title{putoblong draws a rectangle once a canvas is available}
\usage{
putoblong(x1, x2, y1, y2, col = "black", fill = NA, lwd = 1, ...)
}
\arguments{
\item{x1}{defines lefthand edge of rectangle}

\item{x2}{defines right-hand edge or rectangle}

\item{y1}{defines bottom edge of rectangle}

\item{y2}{defines top edge of rectangle}

\item{col}{colour of line. default="black"}

\item{fill}{colour with which to fill the oblong, default=NA = empty}

\item{lwd}{the width of the line, default=1}

\item{...}{to allow for other graphics commands such as lty=2, etc}
}
\value{
an invisible matrix of x,y series defining the polygon
}
\description{
putoblong draws a rectangle after makecanvas has been called
}
\examples{
\dontrun{
   makecanvas(start=0,finish=100)
   putoblong(x1=20,x2=80,y1=25,y2=75,col="red",lwd=2)
}
}
