% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{puttriangle}
\alias{puttriangle}
\title{puttriangle draws a triangle once a canvas is available}
\usage{
puttriangle(xs, ys, col = "black", fill = NA, lwd = 1, ...)
}
\arguments{
\item{xs}{a vector of 3 values defining the x-axis values for the vertices}

\item{ys}{a vector of 3 values defining the y-axis values for the vertices}

\item{col}{the colour of theborder of the triangle}

\item{fill}{the col of the fill of the triangle, default=NA = no fill}

\item{lwd}{the line width if the border is coloured}

\item{...}{to allow for other graphics commands such as lty=2, etc}
}
\value{
an invisible matrix of x,y series defining the polygon
}
\description{
puttriangle draws a triangle once a canvas is available. It
    requires two0 vectors of three values defining the three vertices of the 
    triangle.
}
\examples{
makecanvas()
puttriangle(xs=c(40,60,80),ys=c(40,60,40),col=NA,fill=RGB("red",alpha=127))
}
