% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{replacezeros}
\alias{replacezeros}
\title{replacezeros replaces zeros in columns of a matrix with a given value}
\usage{
replacezeros(x, replacewith = NA)
}
\arguments{
\item{x}{the matrix or data.frame to be processed for zeros}

\item{replacewith}{the replacement value, deafult = 'NA'}
}
\value{
the same input matrix but with zeros replaced with 'replacewith'
}
\description{
replacezeros replaces any zeros that occur in the columns of a
    matrix or data.frame with the value in 'replacewith'. If ther are no 
    zeros then this does nothing. This can be used to remove zeros to avoid
    plots having zeros when really the value is missing.
}
\examples{
x <- matrix(rnorm(25,mean=5,sd=1),nrow=5,ncol=5)
diag(x) <- 0
print(x)
no0x <- replacezeros(x)
print(no0x)
}
