% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{saucompdata}
\alias{saucompdata}
\title{saucompdata plots horizontal barcharts of size-composition data}
\usage{
saucompdata(
  allcomp,
  glb,
  horizline = NULL,
  console = TRUE,
  rundir = "",
  barcol = "red",
  bordercol = "black",
  ylabel = "",
  tabname = ""
)
}
\arguments{
\item{allcomp}{A 3-D array of observed data on size-classes x years x sau}

\item{glb}{the globals objects. specifically glb$nSAU and glb$saunames are
required as a list.}

\item{horizline}{indices of the sizes or ages agsint which to draw reference
lines. default=NULL. IF draw 1 then eg line = 5, if 2 eg = c(5, 13)}

\item{console}{should the graph be plotted to the console or saved as a file.
default=TRUE ie it goes to the console, if set to FALSE it goes to 
paste0(rundir,/)}

\item{rundir}{if saving a png file this is the directory to which a filename
is added}

\item{barcol}{what colour should the bars be, default = red}

\item{bordercol}{what colour should the bar borders be? default = black}

\item{ylabel}{what outer name to be used for the Y-axis, default=''}

\item{tabname}{what web tab name to use. Only useful when using}
}
\value{
nothing but it does generate plots for ech sau
}
\description{
saucompdata generates horizontal plots of all observed size-
    composition data. If there are more than 20 years of data then it will 
    generate two pots of up to 20 years each, separating the plots using
    the plotnumber argument in the file name. This function uses plotcompdata
}
\examples{
print("wait on data")
}
