% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{templateplot}
\alias{templateplot}
\title{templateplot write a plot function template to the console}
\usage{
templateplot()
}
\value{
nothing but does write a template plot to the console
}
\description{
templateplot is used to draft a plotting function that plays 
    well with the makehtml R package. It write the template to the console 
    ready to be copied over to an R file.
}
\examples{
templateplot()
}
