% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot_funs.R
\name{textbox}
\alias{textbox}
\title{textbox draws a rectangle on a canvas adding text is optional}
\usage{
textbox(
  x1,
  x2,
  y1,
  y2,
  col = "black",
  fill = NA,
  lwd = 1,
  txt = "",
  inc = -0.25,
  space = 1,
  ...
)
}
\arguments{
\item{x1}{defines lefthand edge of rectangle}

\item{x2}{defines right-hand edge or rectangle}

\item{y1}{defines bottom edge of rectangle}

\item{y2}{defines top edge of rectangle}

\item{col}{colour of line. default="black"}

\item{fill}{colour with which to fill the oblong, default=NA = empty}

\item{lwd}{the width of the line, default=1}

\item{txt}{default='', = print nothing, the text to be printed centered in
the box. If multiple lines then txt must become a vector.}

\item{inc}{defaukt = -0.25, suitable for cex = 0.85. A number to add to the 
offset from the center point to ensure the text is centered. This might 
need modification if the font size is changed.}

\item{space}{default = 1.0, as with inc, this might need changing if font 
size is changed.}

\item{...}{to allow for other graphics commands such as lty=2, font=2, etc}
}
\value{
a list of the polygon definition of the box and the center coords.
}
\description{
textbox draws a rectangle on a canvas with the option of 
    printing text centered vertically and horizontally inside the box. There
    can be up to four lines of text, but no more.
}
\examples{
canvas <- makecanvas(xstart=0,xfinish=100,ystart=0,yfinish=100,addbox=FALSE)
outbox <- textbox(x1=5,x2=40,y1=90,y2=78,col="black",fill=NA,lwd=1,
                  txt=c("Hello World","A second line","A third line",
                  "A final fourth line, longer than the rest"))
lines(c(5,40),c(outbox[[2]][2],outbox[[2]][2]))
outbox <- textbox(x1=45,x2=95,y1=50,y2=62,col="black",fill=NA,lwd=1,
                  txt=c("Hello World","A second line",
                        "A final third line, longer than the rest"),cex=1.5)
lines(c(45,95),c(outbox[[2]][2],outbox[[2]][2]))
}
