% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot.R
\docType{package}
\name{hplot}
\alias{hplot}
\title{hplot: a set of functions providing aid for case graphics plus custom plots}
\description{
The hplot package provides an array of utility functions and custom graphics
functions.
}
\section{plotutils}{

\describe{
  \item{newplot}{bare-bones plotprep, opens a new device + default par}
  \item{parsyn}{prints par command syntax to the console to be copied}
  \item{plotprep}{sets up a plotting device external to Rstudio}
  \item{parset}{simplifies the definition of main par parameters}
  \item{setplot}{writes a base graphics template to the console}
}
}

