% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hplot.R
\docType{package}
\name{hplot}
\alias{hplot}
\alias{hplot-package}
\title{hplot: a set of functions providing aid for case graphics plus custom plots}
\description{
The hplot package provides an array of utility functions and custom graphics
functions.
}
\seealso{
{
   \link{plot1}, \link{plotprep}, \link{parset}, \link{addnorm}, 
   \link{getmax}, \link{getmin}, \link{pickbound}, \link{yearBubble}, 
   \link{inthist}, \link{plotnull}, \link{plotxyy}, \link{uphist}

}
}
\keyword{internal}
